refactor(@deessejs/next): reorganize src into components/ and lib/ folders
- Move admin-shell.tsx and sidebar-nav.tsx to components/
- Move extract-slug.ts and to-sidebar-items.ts to lib/
- Update internal imports accordingly
- Update package.json exports paths
