+++ packages/next/package.json
@@ -10,12 +10,12 @@
       "types": "./dist/index.d.ts"
     },
     "./admin-shell": {
-      "import": "./dist/admin-shell.js",
-      "types": "./dist/admin-shell.d.ts"
+      "import": "./dist/components/admin-shell.js",
+      "types": "./dist/components/admin-shell.d.ts"
     },
     "./to-sidebar-items": {
-      "import": "./dist/to-sidebar-items.js",
-      "types": "./dist/to-sidebar-items.d.ts"
+      "import": "./dist/lib/to-sidebar-items.js",
+      "types": "./dist/lib/to-sidebar-items.d.ts"
     }
   },
   "scripts": {
