+++ packages/next/src/index.ts
@@ -1,3 +1,5 @@
 // @deessejs/next package
 
 export { RootPage } from "./root-page";
+export { AdminShell } from "./components/admin-shell";
+export { toSidebarItems } from "./lib/to-sidebar-items";
