+++ packages/next/src/lib/to-sidebar-items.ts
